requirements.txt
fastapi
uvicorn
python-dotenv
langchain-community
langchain-google-genai
